mudanças na transferencia de ficheiros

+++ P2P/src/main/java/mensagens/FilePull.java
@@ -6,13 +6,13 @@ import network.Nodo;
 public class FilePull extends Header {
 
     public FileInfo fi;
-    public short[] missingFileChunks;
+    public int[] missingFileChunks;
     public int[] portas;
     public int pps;
 
     public FilePull(){super();}
 
-    public FilePull(String requestID, Nodo origin, FileInfo fi, int[] portas, int pps, short[] missingFileChunks) {
+    public FilePull(String requestID, Nodo origin, FileInfo fi, int[] portas, int pps, int[] missingFileChunks) {
         super(requestID, origin);
         this.fi = fi;
         this.portas = portas;
